requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
asn1crypto==1.5.1
attrs==23.2.0
bcrypt==3.2.2
bidict==0.23.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.10.0
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.7 
dnspython==2.6.1
docstring_parser==0.16
ecdsa==0.19.0
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-mail==1.4.1
fonttools==4.53.0
frozenlist==1.4.1
google-api-core==2.19.1
google-auth==2.30.0
google-cloud==0.34.0
google-cloud-aiplatform==1.59.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.4
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.5
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.0
matplotlib==3.9.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
netifaces==0.10.6
numpy==2.0.0
nylas==6.4.0
openpyxl==3.1.5
orjson==3.10.5
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pg8000==1.31.2
pillow==10.3.0
proto-plus==1.24.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-jose==3.3.0
python-multipart==0.0.9
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
scipy==1.14.0
scramp==1.4.5
shapely==2.0.5
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
validations==0.1.1
watchfiles==0.22.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
